feat: update balance by order id

See merge request NET_09_NguyenThanhDat/cursus-lms!128

+++ Cursus.LMS/Cursus.LMS.Model/DTO/UpsertBalanceDTO.cs
@@ -4,9 +4,9 @@ public class UpsertBalanceDTO
 {
     public string UserId { get; set; }
 
-    public int AvailableBalance { get; set; }
+    public double AvailableBalance { get; set; }
 
-    public int PendingBalance { get; set; }
+    public double PendingBalance { get; set; }
 
     public string Currency { get; set; }
 }
